+++ commands/ping.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-const request = require('request');
+const axios = require('axios');
 
 import { Client } from 'discord.js';
 
@@ -8,47 +8,55 @@ import * as database from '../handlers/database';
 import * as pp from '../handlers/pp';
 
 async function execute(client: Client, msg: any) {
-	var timeDifference : Array<number> = []; 
-	var values : Array<string> = [];
-	var responseStatus = (err : any) => {
-		timeDifference[1] = new Date(Date.now()).getTime();
-		values.push('**OK** (' + (timeDifference[1] - timeDifference[0]) + 'ms)');
-		if (err) values[values.length - 1] = '**FAILED**';
-		timeDifference[0] = new Date(Date.now()).getTime();
+
+	const responsePromise = (request: Promise<any>) => {
+		return new Promise((resolve, reject) => {
+			const initalDate = new Date(Date.now()).getTime();
+			request
+				.then(() => {resolve(new Date(Date.now()).getTime() - initalDate);})
+				.catch((err) => {reject(err);});
+		});
 	};
+
 	const message = await msg.channel.send(':blue_circle: **Pinging**');
 
-	timeDifference[0] = new Date(Date.now()).getTime();
-	request(`https://osu.ppy.sh/api/get_user_recent?${process.env.osuAPI}&u=Moorad`, (err : Error) => {
-		responseStatus(err);
+	var promises = [
+		new Promise((resolve) => {resolve(client.ping)}),
+		responsePromise(axios.get(`https://osu.ppy.sh/api/get_user_recent?k=${process.env.osuAPI}&u=Moorad`)),
+		responsePromise(axios.get('https://the-beautiful-bot-api.herokuapp.com/ping')),
+		responsePromise(new Promise((resovle) => {pp.calculatepp('1192807', {accuracy: 100}, () => {resovle()})})),
+		responsePromise(axios.get('https://api.gatari.pw/user/stats?u=Moorad')),
+		responsePromise(axios.get('https://akatsuki.pw/api/v1/surprise_me')),
+		responsePromise(new Promise((resolve, reject) => {database.read('users', {}, {useCache: false}, (docs, err) => {if (err) reject(); else resolve();});})),
+	];
 
-		request(`https://the-beautiful-bot-api.herokuapp.com/api/player?osuKey=${process.env.osuAPI}&user=Moorad`, (err : Error) => {
-			responseStatus(err);
+	const allSettled = (promise: Promise<any>) => {
+		return promise.then(
+			(value) => {return {status: 'fulfilled', value: value};},
+			(err) => { return {status: 'rejected', reason: err};}
+		);
+	};
 
-			pp.calculatepp('1192807', {
-				accuracy: 100
-			}, () => {
-				responseStatus(err);
+	Promise.all(promises.map(allSettled))
+		.then((values : any) => {
+			var string = '';
 
-				request('https://api.gatari.pw/user/stats?u=Moorad', (err : Error) => {
-					responseStatus(err);
+			database.read('users', {}, {useCache: false}, (docs, err) => {
+				string += 'Discord API: ' + (values[0].status == 'fulfilled' ? `**OK** (${values[0].value ?? ''}ms)` : '**FAILED**');
+				string += '\nosu! API: ' + (values[1].status == 'fulfilled' ? `**OK** (${values[1].value ?? ''}ms)` : '**FAILED**');
+				string += '\nTBB API: ' + (values[2].status == 'fulfilled' ? `**OK** (${values[2].value ?? ''}ms)` : '**FAILED**');
+				string += '\nOjsama: ' + (values[3].status == 'fulfilled' ? `**OK** (${values[3].value ?? ''}ms)` : '**FAILED**');
+				string += '\nGatari: ' + (values[4].status == 'fulfilled' ? `**OK** (${values[4].value ?? ''}ms)` : '**FAILED**');
+				string += '\nAkatsuki: ' + (values[5].status == 'fulfilled' ? `**OK** (${values[5].value ?? ''}ms)` : '**FAILED**');
+				string += '\nDatabase: ' + (values[6].status == 'fulfilled' ? `**OK** (${values[6].value ?? ''}ms)` : '**FAILED**');
+				string += '\n\nNumber of servers: '+ client.guilds.size;
+				string += '\nNumber of users: ' + (!err ? docs.length : '**FAILED**');
 
-					request('https://akatsuki.pw/api/v1/surprise_me', (err : Error) => {
-						responseStatus(err);
-
-						database.read('users', {}, {useCache: false}, (docs, err) => {
-							responseStatus(err);
-
-							message.edit(`Discord API: **OK** (${Math.round(client.ping)}ms)\nosu! API: ${values[0]}\nTBB API: ${values[1]}\nOjsama: ${values[2]}\nGatari: ${values[3]}\nAkatsuki: ${values[4]}\n\nNumber of servers: ${client.guilds.size}\nNumber of users: ${docs.length}`);
-						});
-
-					});
-
-				});
+				message.edit(string);
 			});
 		});
+							
 
-	});
 }
 
 module.exports = {